@@ -1,40 +1,34 @@
-package com.xkcoding.cache.redis.service.impl;
+package com.xkcoding.cache.api;
 
-import com.google.common.collect.Maps;
-import com.xkcoding.cache.redis.entity.User;
-import com.xkcoding.cache.redis.service.UserService;
+import com.xkcoding.cache.entity.User;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.cache.annotation.CachePut;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 
+import java.util.HashMap;
 import java.util.Map;
 
 /**
  * <p>
- * UserService
+ * 模拟用户服务
  * </p>
  *
  * @author yangkai.shen
- * @date Created in 2018-11-15 16:45
+ * @date Created in 2022-09-07 14:08
  */
-@Service
 @Slf4j
-public class UserServiceImpl implements UserService {
+@Service
+public class UserService {
     /**
      * 模拟数据库
      */
-    private static final Map<Long, User> DATABASES = Maps.newConcurrentMap();
-
-    /**
-     * 初始化数据
-     */
-    static {
-        DATABASES.put(1L, new User(1L, "user1"));
-        DATABASES.put(2L, new User(2L, "user2"));
-        DATABASES.put(3L, new User(3L, "user3"));
-    }
+    private static final Map<Long, User> DATABASES = new HashMap<>() {{
+        put(1L, new User(1L, "user1"));
+        put(2L, new User(2L, "user2"));
+        put(3L, new User(3L, "user3"));
+    }};
 
     /**
      * 保存或修改用户
@@ -43,7 +37,6 @@ public class UserServiceImpl implements UserService {
      * @return 操作结果
      */
     @CachePut(value = "user", key = "#user.id")
-    @Override
     public User saveOrUpdate(User user) {
         DATABASES.put(user.getId(), user);
         log.info("保存用户【user】= {}", user);
@@ -57,7 +50,6 @@ public class UserServiceImpl implements UserService {
      * @return 返回结果
      */
     @Cacheable(value = "user", key = "#id")
-    @Override
     public User get(Long id) {
         // 我们假设从数据库读取
         log.info("查询用户【id】= {}", id);
@@ -70,7 +62,6 @@ public class UserServiceImpl implements UserService {
      * @param id key值
      */
     @CacheEvict(value = "user", key = "#id")
-    @Override
     public void delete(Long id) {
         DATABASES.remove(id);
         log.info("删除用户【id】= {}", id);
