@@ -1,23 +1,25 @@
-package com.xkcoding.cache.redis.entity;
+package com.xkcoding.cache.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.io.Serial;
 import java.io.Serializable;
 
 /**
  * <p>
- * 用户实体
+ * 用户对象
  * </p>
  *
  * @author yangkai.shen
- * @date Created in 2018-11-15 16:39
+ * @date Created in 2022-09-07 13:54
  */
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class User implements Serializable {
+    @Serial
     private static final long serialVersionUID = 2892248514883451461L;
     /**
      * 主键id
