@@ -3,10 +3,18 @@ package com.xkcoding.cache.redis;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
+/**
+ * <p>
+ * 启动器
+ * </p>
+ *
+ * @author yangkai.shen
+ * @date Created in 2022-09-06 23:09
+ */
 @SpringBootApplication
-public class SpringBootDemoCacheRedisApplication {
+public class RedisCacheApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(SpringBootDemoCacheRedisApplication.class, args);
+        SpringApplication.run(RedisCacheApplication.class, args);
     }
 }
